misc: fpc1020: Ignore home key presses when touchscreen is in use
When the user's finger is on the touchscreen, don't process any home key
button presses, as an unintended button press will cause poor UX.

Ignore home key presses when the Synaptics s1302 driver is ignoring keypad
presses.

@mcdachpappe
This includes:
* misc: fpc1020: Fix s1302_is_keypad_stopped() logic for home key

There are cases where the home key can be pressed, but when the time comes
to un-press it, s1302_is_keypad_stopped() is true and causes the home key
to be held down. This causes a long home key press to erroneously be
reported, and causes the next home key press to be ignored.

To fix this, make the s1302_is_keypad_stopped() logic only prevent the home
key from being pressed. It will not prevent the home key from being
un-pressed.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: flar2 <asegaert@gmail.com>
Signed-off-by: danny aeng <noreference@web.de>